fix style imports - [CORE-816]
