+++ core/evancore.vue.libs/src/components/dapp-wrapper/mnemonic-export/mnemonic-export.scss
@@ -14,7 +14,7 @@
   the following URL: https://evan.network/license/
 */
 
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .pre-wrap {
   white-space: pre-wrap;
