+++ core/evancore.vue.libs/src/components/failed/failed.scss
@@ -18,7 +18,7 @@
 */
 
 // Colors
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 // Misc
 $curve: cubic-bezier(0.650, 0.000, 0.450, 1.000);
