+++ core/evancore.vue.libs/src/components/forms/checkbox/checkbox.scss
@@ -1,4 +1,4 @@
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .checkbox {
   width: 24px;
