+++ core/evancore.vue.libs/src/components/lists/base-list/base-list.scss
@@ -17,7 +17,7 @@ Fifth Floor, Boston, MA, 02110-1301 USA, or download the license from
 the following URL: https://evan.network/license/
 */
 
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .evan-base-list {
   padding: 0;
