+++ core/evancore.vue.libs/src/components/permissions-editor/permissions-editor.scss
@@ -1,1 +1,1 @@
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
