+++ core/evancore.vue.libs/src/components/permissions/permissions.scss
@@ -1,4 +1,4 @@
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .permissions {
   width: 100%;
