+++ core/evancore.vue.libs/src/components/profile/profile-picture/profile-picture.scss
@@ -17,7 +17,7 @@
   the following URL: https://evan.network/license/
 */
 
-@import "~@evan.network/ui/src/style/utils";
+@import '~@core/ui.libs/src/style/utils';
 
 @mixin masks($size) {
   --size: #{$size};
