+++ core/evancore.vue.libs/src/components/profile/profile-preview/profile-preview.scss
@@ -17,7 +17,7 @@
   the following URL: https://evan.network/license/
 */
 
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .profile-preview {
   &.default {
