+++ core/evancore.vue.libs/src/components/swipe-panel/swipe-panel.scss
@@ -20,7 +20,7 @@
   https://evan.network/license/
 */
 
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .evan-swipe-panel {
   position: relative;
