+++ dapps/assets/src/components/Assets.ts
@@ -19,32 +19,9 @@
 
 // vue imports
 import Component, { mixins } from 'vue-class-component';
-import VueRouter, { Route } from 'vue-router';
 
 // evan.network imports
-import { EvanComponent } from '@evan.network/ui-vue-core';
-
-declare module 'vue/types/vue' {
-  interface Vue {
-    $i18n: any;
-    $router: VueRouter;
-    $route: Route;
-    $store: any;
-    $t: Function;
-    dapp: Dapp;
-    getRuntime: Function;
-  }
-
-  interface Dapp {
-    baseHash: string;
-    baseUrl: string;
-    contractAddress: string;
-    domainName: string;
-    ens: string;
-    fullUrl: string;
-    rootEns: string;
-  }
-}
+import { EvanComponent } from '@core/evancore.vue.libs';
 
 @Component
 export default class AssetsComponent extends mixins(EvanComponent) {
