+++ dapps/assets/src/components/DataContainer.scss
@@ -17,7 +17,7 @@
   the following URL: https://evan.network/license/
 */
 
-@import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .search-hint {
   color: cssVar("gray-600")
