+++ dapps/assets/src/components/digitaltwins/DigitalTwins.scss
@@ -1,4 +1,5 @@
-@import'~@evan.network/ui/src/style/utils';
+// @import '~@evan.network/ui/src/style/utils';
+@import '~@core/ui.libs/src/style/utils';
 
 .content {
   max-width: 850px;
